refactor: modularize attendance service and improve shift validation
- Decompose monolithic attendance.service.ts (1300+ lines) into specialized modules:
  - attendance-processor.ts: Core check-in/out logic and shift resolution
  - attendance-calculator.ts: Metrics and status calculations
  - attendance-automation.ts: Background tasks and auto-checkout
  - attendance-queries.ts: Data retrieval and history
  - attendance-utils.ts: Shared date/time helpers
- Improve shift configuration validation:
  - Add `validateShiftConfig` to detect overlapping times and missing overnight flags
  - Enhance `getBreakError` to correctly handle overnight shift boundaries
  - Fix duration calculations to prevent negative values in non-overnight shifts
  - Ensure break durations do not fallback to legacy defaults when empty
- Update Shifts UI to display real-time timing and break validation errors
